requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.30.0
cryptography==38.0.4
Django==4.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-phonenumber-field==7.0.1
django-twilio==0.14.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-auth==2.15.0
gunicorn==20.1.0
idna==3.4
phonenumbers==8.13.2
Pillow==9.3.0
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.6
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
twilio==7.16.0
urllib3==1.26.13
whitenoise==6.2.0